Fix bugs with moving MDL files to destination
Now if there is no models folder in the place the QC file is located, the folder will be made before moving it so there are no errors for that. Additionally, Snark will now look for any MDL file starting with the name of the mdl file the QC file outputs and ends with the .mdl extension so any other mdl files needed (e.g. BarneyT.mdl) are found and moved.